requirements.txt
py-cord==2.4.0
tortoise-orm==0.19.3
python-dotenv==1.0.0
black==23.3.0
pytest==7.3.1
pytest-asyncio==0.21.0